frontend: move default interface states to /constants
