+++ frontend/src/constants/defaultErrorData.ts
@@ -1,0 +1,7 @@
+import { IErrorData } from "../interfaces/IErrorData";
+
+export const defaultErrorData: IErrorData = {
+  error: "no error",
+  message: "no error",
+  statusCode: 0,
+};
