+++ frontend/src/constants/defaultErrorState.ts
@@ -1,0 +1,7 @@
+import { defaultErrorData } from "../constants/defaultErrorData";
+import { IErrorContext } from "../interfaces/IErrorContext";
+
+export const defaultErrorState: IErrorContext = {
+  errorData: defaultErrorData,
+  showError: false,
+};
