+++ frontend/src/interfaces/IErrorData.ts
@@ -3,9 +3,3 @@ export interface IErrorData {
   message: string;
   statusCode: number;
 }
-
-export const defaultErrorData: IErrorData = {
-  error: "no error",
-  message: "no error",
-  statusCode: 0,
-};
