+++ frontend/src/interfaces/ISocketContext.ts
@@ -1,15 +1,5 @@
-import { io, Socket } from "socket.io-client";
-
-import getConfig from "next/config";
-const { publicRuntimeConfig } = getConfig();
+import { Socket } from "socket.io-client";
 
 export interface ISocketContext {
   socket: Socket;
 }
-
-export const defaultSocketState = {
-  socket: io(
-    `http://${publicRuntimeConfig.HOST}:${publicRuntimeConfig.WEBSOCKETS_PORT}`,
-    { transports: ["websocket"] }
-  ),
-};
